requirements.txt
boto3
discord.py
jsonschema
openai
pyyaml